+++ CHANGELOG.rst
@@ -1,10 +1,13 @@
 Changelog
 =========
 
+
 1.1.6 (2016-08-04)
 ------------------
 
 * background style for text in header images
+
+* nicer footer
 
 
 1.1.5 (2016-07-03)